+++ product-microservice/src/main/java/product/microservice/productmicroservice/repository/CustomerRepository.java
@@ -3,5 +3,5 @@ package product.microservice.productmicroservice.repository;
 import org.springframework.data.repository.CrudRepository;
 import product.microservice.productmicroservice.model.Customer;
 
-public interface CustomerRepository extends CrudRepository<Customer, Integer> {
+public interface CustomerRepository extends CrudRepository<Customer, Long> {
 }
