+++ product-microservice/src/main/java/product/microservice/productmicroservice/repository/ReviewRepository.java
@@ -3,5 +3,5 @@ package product.microservice.productmicroservice.repository;
 import org.springframework.data.repository.CrudRepository;
 import product.microservice.productmicroservice.model.Review;
 
-public interface ReviewRepository extends CrudRepository<Review, Integer> {
+public interface ReviewRepository extends CrudRepository<Review, Long> {
 }
